app.py
from flask import Flask, request, jsonify, render_template
import cv2
import numpy as np
from PIL import Image
import io
import json

app = Flask(__name__)

# Security vulnerability detection functions

def detect_perimeter_vulnerabilities(image):
    # This is a placeholder function that would implement actual detection logic
    # In a real implementation, this would use computer vision techniques to:
    # - Detect fence/gate integrity
    # - Identify gaps in perimeter
    # - Check for unauthorized access points
    vulnerabilities = []
    # Add mock detection results
    vulnerabilities.append({
        "type": "Perimeter Gap",
        "location": "(100, 200)",
        "severity": "High",
        "description": "Potential gap in perimeter fence"
    })
    return vulnerabilities

def detect_camera_coverage(image):
    # This would implement actual camera coverage analysis
    vulnerabilities = []
    # Add mock detection results
    vulnerabilities.append({
        "type": "Camera Blind Spot",
        "location": "(300, 400)",
        "severity": "Medium",
        "description": "Area not covered by security cameras"
    })
    return vulnerabilities

def analyze_image(image_bytes):
    # Convert bytes to OpenCV image
    image = cv2.imdecode(np.frombuffer(image_bytes, np.uint8), cv2.IMREAD_COLOR)
    
    # Run different detection algorithms
    perimeter_vulns = detect_perimeter_vulnerabilities(image)
    camera_vulns = detect_camera_coverage(image)
    
    # Combine results
    all_vulnerabilities = {
        "perimeter_vulnerabilities": perimeter_vulns,
        "camera_coverage_vulnerabilities": camera_vulns,
        "total_vulnerabilities": len(perimeter_vulns) + len(camera_vulns)
    }
    
    return all_vulnerabilities

@app.route('/')
def index():
    return render_template('index.html')

@app.route('/api/analyze', methods=['POST'])
def analyze_photo():
    if 'image' not in request.files:
        return jsonify({'error': 'No image provided'}), 400
    
    image_file = request.files['image']
    image_bytes = image_file.read()
    
    try:
        results = analyze_image(image_bytes)
        return jsonify(results)
    except Exception as e:
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000)
