requirements.txt
flask==3.0.0
opencv-python==4.8.0.74
numpy==1.24.3
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
